move pygame functionality to dedicated module bapsf_motion.gui.configure.pygame_
